requirements.txt
Flask==2.3.3                # Framework web principal
Flask-SQLAlchemy==3.0.5     # ORM pour gérer la base de données
Flask-JWT-Extended==4.4.4   # Pour gérer l'authentification via JWT
Flask-Migrate               # Pour gérer les migrations

psycopg2-binary==2.9.11     # Si tu passes à PostgreSQL plus tard

marshmallow==3.22.0         # Pour sérialisation/désérialisation JSON
python-dotenv==1.0.0        # Pour charger variables d'environnement

pytest==8.3.3                # Pour tests unitaires
pytest-flask==1.3.0          # Pour tester tes endpoints Flask
flask-cors==3.0.1            # Pour autoriser les requêtes frontend

pandas                       # Pour lire un document excel
openpyxl